Move layout related components to components/layout/
